backend/requirements.txt
# Core
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3

# Redis
redis[hiredis]==5.0.1

# Validation & Settings
pydantic==2.6.1
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# HTTP & Networking
httpx==0.26.0

# Rate Limiting
slowapi==0.1.9

# Utilities
python-dateutil==2.8.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0
factory-boy==3.3.0
